Limit stellar visibility to luminosity box.

+++ src/classes/celestial.h
@@ -58,7 +58,7 @@ class CelestialObject
     cel_obj_type type = star;
     std::string name;
 
-    int drawnx=-1e9, drawny=-1e9;
+    float drawnx=-1e9, drawny=-1e9;
 
     CelestialObject() {};
     CelestialLocation location;
